# openapi.yaml
openapi: 3.1.0
info:
  title: RescueBite API
  description: |
    Food Waste Reduction Marketplace REST API

    Built with FastAPI + SQLModel + PostgreSQL + Redis + Celery.

    **Auth flow:** Register → Verify Email (OTP) → Login → Bearer token → Refresh → Logout

    **Business flow:**
    1. Vendor registers and submits business profile
    2. Admin approves vendor
    3. Vendor creates listings (food items with discount)
    4. Customer browses listings, places order
    5. Customer pays via Kaspi Pay (or simulated)
    6. Vendor confirms and marks order ready for pickup
    7. Customer picks up with token
    8. Celery Beat applies price decay every 72h
  version: 1.0.0
  contact:
    name: RescueBite Team
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /vendors/register:
    post:
      tags: [Vendors]
      summary: Register vendor profile (vendor role required)
      operationId: registerVendor
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [business_name, bin_number, address, latitude, longitude]
              properties:
                business_name: { type: string, example: "Сushi Бар Алматы" }
                bin_number: { type: string, example: "123456789012" }
                address: { type: string, example: "ул. Абая 150, Алматы" }
                latitude: { type: number, example: 43.238 }
                longitude: { type: number, example: 76.945 }
      responses:
        "201":
          description: Vendor profile created (pending approval)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorResponse' }
        "409": { $ref: '#/components/responses/Conflict' }
  /vendors/me:
    get:
      tags: [Vendors]
      summary: Get own vendor profile
      operationId: getMyVendor
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
  /vendors/{vendor_id}:
    get:
      tags: [Vendors]
      summary: Get vendor by ID (public)
      operationId: getVendor
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Vendor details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VendorResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
  /payments/{order_id}/initiate:
    post:
      tags: [Payments]
      summary: Initiate Kaspi Pay payment for an order
      operationId: initiatePayment
      security: [{ BearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment initiated with Kaspi URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentInitResponse' }
        "409": { $ref: '#/components/responses/Conflict' }
  /payments/{order_id}/simulate-success:
    post:
      tags: [Payments]
      summary: "[DEV] Simulate successful Kaspi payment"
      operationId: simulatePayment
      security: [{ BearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment simulated, order confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentStatusResponse' }
        "409": { $ref: '#/components/responses/Conflict' }
  /payments/{order_id}/status:
    get:
      tags: [Payments]
      summary: Get payment status for an order
      operationId: getPaymentStatus
      security: [{ BearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentStatusResponse' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /admin/vendors/{vendor_id}:
    delete:
      tags: [Admin]
      summary: Hard-delete a vendor and all their listings
      operationId: adminDeleteVendor
      security: [{ BearerAuth: [] }]
      parameters:
        - name: vendor_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Vendor deleted
        "409": { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
    Conflict:
      description: Conflict — duplicate resource or business rule violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorDetail' }
...
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail: { type: string }
...
    VendorResponse:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        business_name: { type: string }
        bin_number: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        is_approved: { type: boolean }
    PaymentInitResponse:
      type: object
      properties:
        order_id: { type: integer }
        amount_kzt: { type: integer }
        kaspi_payment_url: { type: string }
        payment_id: { type: integer }
        status: { type: string }
    PaymentStatusResponse:
      type: object
      properties:
        payment_id: { type: integer }
        order_id: { type: integer }
        amount_kzt: { type: integer }
        status: { type: string }
        kaspi_ref_id: { type: string, nullable: true }
        paid_at: { type: string, format: date-time, nullable: true }
...